requirements.txt
pynput==1.7.6
mss==9.0.1
groq>=0.12.0
google-generativeai>=0.3.0
Pillow==10.3.0
python-dotenv==1.0.0
